# apis/gateway-api/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway API
  version: 1.0.0
  description: Public API gateway exposing core shipment lifecycle and auth.
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging-api.example.com
    description: Staging
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: Invalid request }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
    Address:
      type: object
      required: [name, line1, city, state, postalCode, country]
      properties:
        name: { type: string }
        line1: { type: string, maxLength: 128 }
        line2: { type: string, maxLength: 128 }
        city: { type: string }
        state: { type: string, minLength: 2, maxLength: 64 }
        postalCode: { type: string, maxLength: 16 }
        country: { type: string, minLength: 2, maxLength: 2, example: US }
    Package:
      type: object
      required: [weightKg, lengthCm, widthCm, heightCm]
      properties:
        weightKg: { type: number, minimum: 0.01 }
        lengthCm: { type: number, minimum: 1 }
        widthCm: { type: number, minimum: 1 }
        heightCm: { type: number, minimum: 1 }
        declaredValue: { type: number, minimum: 0 }
        commodityType: { type: string, example: general }
        hazmat: { type: boolean, default: false }
    Shipment:
      type: object
      required: [id, trackingNumber, status]
      properties:
        id: { type: string, format: uuid }
        trackingNumber: { type: string }
        status: { type: string, enum: [CREATED, READY_FOR_PICKUP, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, EXCEPTION] }
        serviceLevel: { type: string, enum: [GROUND, TWO_DAY, OVERNIGHT, SAME_DAY] }
        shipperId: { type: string }
        from: { $ref: '#/components/schemas/Address' }
        to: { $ref: '#/components/schemas/Address' }
        packages:
          type: array
          items: { $ref: '#/components/schemas/Package' }
        createdAt: { type: string, format: date-time }
    ShipmentCreateRequest:
      type: object
      required: [serviceLevel, from, to, packages]
      properties:
        serviceLevel: { type: string, enum: [GROUND, TWO_DAY, OVERNIGHT, SAME_DAY] }
        from: { $ref: '#/components/schemas/Address' }
        to: { $ref: '#/components/schemas/Address' }
        packages:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Package' }
      example:
        serviceLevel: TWO_DAY
        from: { name: Sender, line1: 1 Main St, city: Austin, state: TX, postalCode: 78701, country: US }
        to: { name: Receiver, line1: 99 Market, city: San Francisco, state: CA, postalCode: 94103, country: US }
        packages: [{ weightKg: 2.5, lengthCm: 30, widthCm: 20, heightCm: 10 }]
...
paths:
...
  /v1/shipments:
    post:
      summary: Create a shipment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShipmentCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Shipment' }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/shipments/{id}:
    get:
      summary: Get a shipment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Shipment' } } } }
        '404': { description: Not Found }
    patch:
      summary: Update shipment metadata (pre-pickup)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                serviceLevel: { type: string }
                to: { $ref: '#/components/schemas/Address' }
      responses:
        '200': { description: OK }
        '400': { description: Validation error }
...
  /v1/track/{trackingNumber}:
    get:
      summary: Public tracking endpoint
      security: []
      parameters:
        - in: path
          name: trackingNumber
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  trackingNumber: { type: string }
                  status: { type: string }
                  eta: { type: string, format: date-time }
                  events:
                    type: array
                    items:
                      type: object
                      properties:
                        type: { type: string }
                        ts: { type: string, format: date-time }
                        hub: { type: string }
        '404': { description: Not Found }
...